Refectored Rust code
