+++ src/native/src/core/dictionary.rs
